Move AzleProgram impls around
Trying to break out into multiple impls, each to be located with
corresponding code.
