+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/inspect_message/mod.rs
@@ -3,8 +3,8 @@ use cdk_framework::act::node::{
 };
 
 use crate::{
-    canister_method::errors,
-    ts_ast::{azle_program::HelperMethods, TsAst},
+    canister_method::{errors, get_azle_fn_decls::GetProgramAzleFnDecls},
+    TsAst,
 };
 
 mod rust;
