+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/post_upgrade/mod.rs
@@ -1,8 +1,8 @@
 use cdk_framework::act::node::canister_method::{CanisterMethodType, PostUpgradeMethod};
 
 use crate::{
-    canister_method::errors,
-    ts_ast::{azle_program::HelperMethods, TsAst},
+    canister_method::{errors, get_azle_fn_decls::GetProgramAzleFnDecls},
+    TsAst,
 };
 
 mod rust;
